+++ custom_components/myskoda/utils.py
@@ -2,15 +2,10 @@ import logging
 
 from typing import Callable
 
-from aiohttp.client_exceptions import ClientResponseError
-from homeassistant.config_entries import ConfigEntry
-from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import UpdateFailed
 
 from .coordinator import MySkodaDataUpdateCoordinator
 from .entity import MySkodaEntity
-from .issues import async_create_spin_issue
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -32,28 +27,3 @@ def add_supported_entities(
                     entities.append(sensor)
 
     async_add_entities(entities, update_before_add=True)
-
-
-def handle_aiohttp_error(
-    poll_type: str,
-    e: ClientResponseError,
-    hass: HomeAssistant,
-    config: ConfigEntry,
-) -> None:
-    _LOGGER.debug("Received error %d with content %s", e.status, e.message)
-
-    if e.status == 412:
-        # Handle precondition failed by creating an issue for incorrect S-PIN
-        async_create_spin_issue(hass, config.entry_id)
-        return
-
-    elif e.status == 500:
-        # Log message for error 500, otherwise ignore
-        _LOGGER.error(
-            f"Error requesting {poll_type} from MySkoda API: {e.message} ({e.status}), ignoring this"
-        )
-        return
-    else:
-        raise UpdateFailed(
-            f"Error requesting {poll_type} from MySkoda API: {e.message} ({e.status})"
-        )
